race-ms/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.0
h11==0.14.0
idna==3.4
pydantic==1.10.7
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
fastapi==0.95.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pydantic==1.10.7
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2022.7.1
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
aiosqlite==0.17.0
anyio==3.6.2
asyncpg==0.27.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pyasn1==0.4.8
pydantic==1.10.7
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
websockets==10.4
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pyasn1==0.4.8
pydantic==1.10.7
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
redis==4.5.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
websockets==10.4
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
click==8.1.3
ecdsa==0.18.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
iso8601==1.1.0
pyasn1==0.4.8
pydantic==1.10.7
PyJWT==2.6.0
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
redis==4.5.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
tortoise-orm==0.19.3
typing_extensions==4.5.0
uvicorn==0.21.1
websockets==10.4
